+++ vsc7448-pac/src/ana_ac.rs
@@ -191,9 +191,9 @@ impl PS_COMMON {
         RegisterAddress::new(self.0 + 0x1c)
     }
     #[inline(always)]
-    pub fn VSTAX_CTRL(&self, index: u32) -> RegisterAddress<ps_common::VSTAX_CTRL> {
+    pub fn VSTAX_CTRL(&self, index: u8) -> RegisterAddress<ps_common::VSTAX_CTRL> {
         assert!(index < 53);
-        RegisterAddress::new(self.0 + 0x30 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x30 + u32::from(index) * 0x4)
     }
     #[inline(always)]
     pub fn VSTAX_GMIRROR_CFG(&self) -> RegisterAddress<ps_common::VSTAX_GMIRROR_CFG> {
@@ -262,14 +262,14 @@ impl SRC {
 pub struct STAT_CNT_CFG_ACL(pub(super) u32);
 impl STAT_CNT_CFG_ACL {
     #[inline(always)]
-    pub fn STAT_LSB_CNT(&self, index: u32) -> RegisterAddress<stat_cnt_cfg_acl::STAT_LSB_CNT> {
+    pub fn STAT_LSB_CNT(&self, index: u8) -> RegisterAddress<stat_cnt_cfg_acl::STAT_LSB_CNT> {
         assert!(index < 2);
-        RegisterAddress::new(self.0 + 0x0 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x0 + u32::from(index) * 0x4)
     }
     #[inline(always)]
-    pub fn STAT_MSB_CNT(&self, index: u32) -> RegisterAddress<stat_cnt_cfg_acl::STAT_MSB_CNT> {
+    pub fn STAT_MSB_CNT(&self, index: u8) -> RegisterAddress<stat_cnt_cfg_acl::STAT_MSB_CNT> {
         assert!(index < 2);
-        RegisterAddress::new(self.0 + 0x8 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x8 + u32::from(index) * 0x4)
     }
 }
 
@@ -277,14 +277,14 @@ impl STAT_CNT_CFG_ACL {
 pub struct STAT_CNT_CFG_BDLB(pub(super) u32);
 impl STAT_CNT_CFG_BDLB {
     #[inline(always)]
-    pub fn STAT_LSB_CNT(&self, index: u32) -> RegisterAddress<stat_cnt_cfg_bdlb::STAT_LSB_CNT> {
+    pub fn STAT_LSB_CNT(&self, index: u8) -> RegisterAddress<stat_cnt_cfg_bdlb::STAT_LSB_CNT> {
         assert!(index < 2);
-        RegisterAddress::new(self.0 + 0x0 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x0 + u32::from(index) * 0x4)
     }
     #[inline(always)]
-    pub fn STAT_MSB_CNT(&self, index: u32) -> RegisterAddress<stat_cnt_cfg_bdlb::STAT_MSB_CNT> {
+    pub fn STAT_MSB_CNT(&self, index: u8) -> RegisterAddress<stat_cnt_cfg_bdlb::STAT_MSB_CNT> {
         assert!(index < 2);
-        RegisterAddress::new(self.0 + 0x8 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x8 + u32::from(index) * 0x4)
     }
 }
 
@@ -292,14 +292,14 @@ impl STAT_CNT_CFG_BDLB {
 pub struct STAT_CNT_CFG_BUM(pub(super) u32);
 impl STAT_CNT_CFG_BUM {
     #[inline(always)]
-    pub fn STAT_LSB_CNT(&self, index: u32) -> RegisterAddress<stat_cnt_cfg_bum::STAT_LSB_CNT> {
+    pub fn STAT_LSB_CNT(&self, index: u8) -> RegisterAddress<stat_cnt_cfg_bum::STAT_LSB_CNT> {
         assert!(index < 6);
-        RegisterAddress::new(self.0 + 0x0 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x0 + u32::from(index) * 0x4)
     }
     #[inline(always)]
-    pub fn STAT_MSB_CNT(&self, index: u32) -> RegisterAddress<stat_cnt_cfg_bum::STAT_MSB_CNT> {
+    pub fn STAT_MSB_CNT(&self, index: u8) -> RegisterAddress<stat_cnt_cfg_bum::STAT_MSB_CNT> {
         assert!(index < 6);
-        RegisterAddress::new(self.0 + 0x18 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x18 + u32::from(index) * 0x4)
     }
 }
 
@@ -307,14 +307,14 @@ impl STAT_CNT_CFG_BUM {
 pub struct STAT_CNT_CFG_ERLEG(pub(super) u32);
 impl STAT_CNT_CFG_ERLEG {
     #[inline(always)]
-    pub fn STAT_LSB_CNT(&self, index: u32) -> RegisterAddress<stat_cnt_cfg_erleg::STAT_LSB_CNT> {
+    pub fn STAT_LSB_CNT(&self, index: u8) -> RegisterAddress<stat_cnt_cfg_erleg::STAT_LSB_CNT> {
         assert!(index < 8);
-        RegisterAddress::new(self.0 + 0x0 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x0 + u32::from(index) * 0x4)
     }
     #[inline(always)]
-    pub fn STAT_MSB_CNT(&self, index: u32) -> RegisterAddress<stat_cnt_cfg_erleg::STAT_MSB_CNT> {
+    pub fn STAT_MSB_CNT(&self, index: u8) -> RegisterAddress<stat_cnt_cfg_erleg::STAT_MSB_CNT> {
         assert!(index < 8);
-        RegisterAddress::new(self.0 + 0x20 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x20 + u32::from(index) * 0x4)
     }
 }
 
@@ -322,14 +322,14 @@ impl STAT_CNT_CFG_ERLEG {
 pub struct STAT_CNT_CFG_IRLEG(pub(super) u32);
 impl STAT_CNT_CFG_IRLEG {
     #[inline(always)]
-    pub fn STAT_LSB_CNT(&self, index: u32) -> RegisterAddress<stat_cnt_cfg_irleg::STAT_LSB_CNT> {
+    pub fn STAT_LSB_CNT(&self, index: u8) -> RegisterAddress<stat_cnt_cfg_irleg::STAT_LSB_CNT> {
         assert!(index < 8);
-        RegisterAddress::new(self.0 + 0x0 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x0 + u32::from(index) * 0x4)
     }
     #[inline(always)]
-    pub fn STAT_MSB_CNT(&self, index: u32) -> RegisterAddress<stat_cnt_cfg_irleg::STAT_MSB_CNT> {
+    pub fn STAT_MSB_CNT(&self, index: u8) -> RegisterAddress<stat_cnt_cfg_irleg::STAT_MSB_CNT> {
         assert!(index < 8);
-        RegisterAddress::new(self.0 + 0x20 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x20 + u32::from(index) * 0x4)
     }
 }
 
@@ -337,14 +337,14 @@ impl STAT_CNT_CFG_IRLEG {
 pub struct STAT_CNT_CFG_ISDX(pub(super) u32);
 impl STAT_CNT_CFG_ISDX {
     #[inline(always)]
-    pub fn STAT_LSB_CNT(&self, index: u32) -> RegisterAddress<stat_cnt_cfg_isdx::STAT_LSB_CNT> {
+    pub fn STAT_LSB_CNT(&self, index: u8) -> RegisterAddress<stat_cnt_cfg_isdx::STAT_LSB_CNT> {
         assert!(index < 6);
-        RegisterAddress::new(self.0 + 0x0 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x0 + u32::from(index) * 0x4)
     }
     #[inline(always)]
-    pub fn STAT_MSB_CNT(&self, index: u32) -> RegisterAddress<stat_cnt_cfg_isdx::STAT_MSB_CNT> {
+    pub fn STAT_MSB_CNT(&self, index: u8) -> RegisterAddress<stat_cnt_cfg_isdx::STAT_MSB_CNT> {
         assert!(index < 3);
-        RegisterAddress::new(self.0 + 0x18 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x18 + u32::from(index) * 0x4)
     }
 }
 
@@ -352,23 +352,23 @@ impl STAT_CNT_CFG_ISDX {
 pub struct STAT_CNT_CFG_PORT(pub(super) u32);
 impl STAT_CNT_CFG_PORT {
     #[inline(always)]
-    pub fn STAT_CFG(&self, index: u32) -> RegisterAddress<stat_cnt_cfg_port::STAT_CFG> {
+    pub fn STAT_CFG(&self, index: u8) -> RegisterAddress<stat_cnt_cfg_port::STAT_CFG> {
         assert!(index < 4);
-        RegisterAddress::new(self.0 + 0x4 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x4 + u32::from(index) * 0x4)
     }
     #[inline(always)]
     pub fn STAT_EVENTS_STICKY(&self) -> RegisterAddress<stat_cnt_cfg_port::STAT_EVENTS_STICKY> {
         RegisterAddress::new(self.0 + 0x0)
     }
     #[inline(always)]
-    pub fn STAT_LSB_CNT(&self, index: u32) -> RegisterAddress<stat_cnt_cfg_port::STAT_LSB_CNT> {
+    pub fn STAT_LSB_CNT(&self, index: u8) -> RegisterAddress<stat_cnt_cfg_port::STAT_LSB_CNT> {
         assert!(index < 4);
-        RegisterAddress::new(self.0 + 0x14 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x14 + u32::from(index) * 0x4)
     }
     #[inline(always)]
-    pub fn STAT_MSB_CNT(&self, index: u32) -> RegisterAddress<stat_cnt_cfg_port::STAT_MSB_CNT> {
+    pub fn STAT_MSB_CNT(&self, index: u8) -> RegisterAddress<stat_cnt_cfg_port::STAT_MSB_CNT> {
         assert!(index < 4);
-        RegisterAddress::new(self.0 + 0x24 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x24 + u32::from(index) * 0x4)
     }
 }
 
@@ -376,14 +376,14 @@ impl STAT_CNT_CFG_PORT {
 pub struct STAT_CNT_CFG_QUEUE(pub(super) u32);
 impl STAT_CNT_CFG_QUEUE {
     #[inline(always)]
-    pub fn STAT_LSB_CNT(&self, index: u32) -> RegisterAddress<stat_cnt_cfg_queue::STAT_LSB_CNT> {
+    pub fn STAT_LSB_CNT(&self, index: u8) -> RegisterAddress<stat_cnt_cfg_queue::STAT_LSB_CNT> {
         assert!(index < 2);
-        RegisterAddress::new(self.0 + 0x0 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x0 + u32::from(index) * 0x4)
     }
     #[inline(always)]
-    pub fn STAT_MSB_CNT(&self, index: u32) -> RegisterAddress<stat_cnt_cfg_queue::STAT_MSB_CNT> {
+    pub fn STAT_MSB_CNT(&self, index: u8) -> RegisterAddress<stat_cnt_cfg_queue::STAT_MSB_CNT> {
         assert!(index < 2);
-        RegisterAddress::new(self.0 + 0x8 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x8 + u32::from(index) * 0x4)
     }
 }
 
@@ -393,26 +393,26 @@ impl STAT_GLOBAL_CFG_ACL {
     #[inline(always)]
     pub fn GLOBAL_CNT_FRM_TYPE_CFG(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_acl::GLOBAL_CNT_FRM_TYPE_CFG> {
         assert!(index < 2);
-        RegisterAddress::new(self.0 + 0x0 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x0 + u32::from(index) * 0x4)
     }
     #[inline(always)]
     pub fn STAT_GLOBAL_CFG(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_acl::STAT_GLOBAL_CFG> {
         assert!(index < 2);
-        RegisterAddress::new(self.0 + 0x8 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x8 + u32::from(index) * 0x4)
     }
     #[inline(always)]
     pub fn STAT_GLOBAL_EVENT_MASK(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_acl::STAT_GLOBAL_EVENT_MASK> {
         assert!(index < 2);
-        RegisterAddress::new(self.0 + 0x10 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x10 + u32::from(index) * 0x4)
     }
 }
 
@@ -422,26 +422,26 @@ impl STAT_GLOBAL_CFG_BDLB {
     #[inline(always)]
     pub fn GLOBAL_CNT_FRM_TYPE_CFG(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_bdlb::GLOBAL_CNT_FRM_TYPE_CFG> {
         assert!(index < 2);
-        RegisterAddress::new(self.0 + 0x0 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x0 + u32::from(index) * 0x4)
     }
     #[inline(always)]
     pub fn STAT_GLOBAL_CFG(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_bdlb::STAT_GLOBAL_CFG> {
         assert!(index < 2);
-        RegisterAddress::new(self.0 + 0x8 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x8 + u32::from(index) * 0x4)
     }
     #[inline(always)]
     pub fn STAT_GLOBAL_EVENT_MASK(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_bdlb::STAT_GLOBAL_EVENT_MASK> {
         assert!(index < 2);
-        RegisterAddress::new(self.0 + 0x10 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x10 + u32::from(index) * 0x4)
     }
 }
 
@@ -451,26 +451,26 @@ impl STAT_GLOBAL_CFG_BUM {
     #[inline(always)]
     pub fn GLOBAL_CNT_FRM_TYPE_CFG(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_bum::GLOBAL_CNT_FRM_TYPE_CFG> {
         assert!(index < 6);
-        RegisterAddress::new(self.0 + 0x0 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x0 + u32::from(index) * 0x4)
     }
     #[inline(always)]
     pub fn STAT_GLOBAL_CFG(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_bum::STAT_GLOBAL_CFG> {
         assert!(index < 6);
-        RegisterAddress::new(self.0 + 0x18 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x18 + u32::from(index) * 0x4)
     }
     #[inline(always)]
     pub fn STAT_GLOBAL_EVENT_MASK(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_bum::STAT_GLOBAL_EVENT_MASK> {
         assert!(index < 6);
-        RegisterAddress::new(self.0 + 0x30 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x30 + u32::from(index) * 0x4)
     }
 }
 
@@ -480,26 +480,26 @@ impl STAT_GLOBAL_CFG_ERLEG {
     #[inline(always)]
     pub fn GLOBAL_CNT_FRM_TYPE_CFG(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_erleg::GLOBAL_CNT_FRM_TYPE_CFG> {
         assert!(index < 8);
-        RegisterAddress::new(self.0 + 0x0 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x0 + u32::from(index) * 0x4)
     }
     #[inline(always)]
     pub fn STAT_GLOBAL_CFG(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_erleg::STAT_GLOBAL_CFG> {
         assert!(index < 8);
-        RegisterAddress::new(self.0 + 0x20 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x20 + u32::from(index) * 0x4)
     }
     #[inline(always)]
     pub fn STAT_GLOBAL_EVENT_MASK(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_erleg::STAT_GLOBAL_EVENT_MASK> {
         assert!(index < 8);
-        RegisterAddress::new(self.0 + 0x40 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x40 + u32::from(index) * 0x4)
     }
 }
 
@@ -509,26 +509,26 @@ impl STAT_GLOBAL_CFG_IRLEG {
     #[inline(always)]
     pub fn GLOBAL_CNT_FRM_TYPE_CFG(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_irleg::GLOBAL_CNT_FRM_TYPE_CFG> {
         assert!(index < 8);
-        RegisterAddress::new(self.0 + 0x0 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x0 + u32::from(index) * 0x4)
     }
     #[inline(always)]
     pub fn STAT_GLOBAL_CFG(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_irleg::STAT_GLOBAL_CFG> {
         assert!(index < 8);
-        RegisterAddress::new(self.0 + 0x20 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x20 + u32::from(index) * 0x4)
     }
     #[inline(always)]
     pub fn STAT_GLOBAL_EVENT_MASK(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_irleg::STAT_GLOBAL_EVENT_MASK> {
         assert!(index < 8);
-        RegisterAddress::new(self.0 + 0x40 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x40 + u32::from(index) * 0x4)
     }
 }
 
@@ -538,26 +538,26 @@ impl STAT_GLOBAL_CFG_ISDX {
     #[inline(always)]
     pub fn GLOBAL_CNT_FRM_TYPE_CFG(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_isdx::GLOBAL_CNT_FRM_TYPE_CFG> {
         assert!(index < 6);
-        RegisterAddress::new(self.0 + 0x0 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x0 + u32::from(index) * 0x4)
     }
     #[inline(always)]
     pub fn STAT_GLOBAL_CFG(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_isdx::STAT_GLOBAL_CFG> {
         assert!(index < 6);
-        RegisterAddress::new(self.0 + 0x18 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x18 + u32::from(index) * 0x4)
     }
     #[inline(always)]
     pub fn STAT_GLOBAL_EVENT_MASK(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_isdx::STAT_GLOBAL_EVENT_MASK> {
         assert!(index < 6);
-        RegisterAddress::new(self.0 + 0x30 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x30 + u32::from(index) * 0x4)
     }
 }
 
@@ -567,10 +567,10 @@ impl STAT_GLOBAL_CFG_PORT {
     #[inline(always)]
     pub fn STAT_GLOBAL_EVENT_MASK(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_port::STAT_GLOBAL_EVENT_MASK> {
         assert!(index < 4);
-        RegisterAddress::new(self.0 + 0x0 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x0 + u32::from(index) * 0x4)
     }
     #[inline(always)]
     pub fn STAT_RESET(&self) -> RegisterAddress<stat_global_cfg_port::STAT_RESET> {
@@ -584,26 +584,26 @@ impl STAT_GLOBAL_CFG_QUEUE {
     #[inline(always)]
     pub fn GLOBAL_CNT_FRM_TYPE_CFG(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_queue::GLOBAL_CNT_FRM_TYPE_CFG> {
         assert!(index < 2);
-        RegisterAddress::new(self.0 + 0x0 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x0 + u32::from(index) * 0x4)
     }
     #[inline(always)]
     pub fn STAT_GLOBAL_CFG(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_queue::STAT_GLOBAL_CFG> {
         assert!(index < 2);
-        RegisterAddress::new(self.0 + 0x8 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x8 + u32::from(index) * 0x4)
     }
     #[inline(always)]
     pub fn STAT_GLOBAL_EVENT_MASK(
         &self,
-        index: u32,
+        index: u8,
     ) -> RegisterAddress<stat_global_cfg_queue::STAT_GLOBAL_EVENT_MASK> {
         assert!(index < 2);
-        RegisterAddress::new(self.0 + 0x10 + index * 0x4)
+        RegisterAddress::new(self.0 + 0x10 + u32::from(index) * 0x4)
     }
 }
